+++ lib/tusc/offset_response.rb
@@ -6,4 +6,16 @@ class TusClient::OffsetResponse
   def offset
     @response.header['Upload-Offset'].to_i # nil.to_i == 0
   end
+
+  def raw
+    @response
+  end
+
+  def status_code
+    @response.code.to_i
+  end
+
+  def success?
+    status_code == 204 && (offset >= 0)
+  end
 end