+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Upcoming v0.4.3: FIX: Can upload video files. Testing via tus-server.
 - Added tests against local tus-server
 - FIX: Uploader now supports uploading video files (using octet-stream vs. detected content type)
   - removes dependency on MimeMagic
+- All *Response objects have basic interface: raw, status_code, success?
 
 2020-07-23 v0.4.2: FIX: Uploader#push_chunk had error logging (truncated) body